+++ core/interfaces/core.go
@@ -27,16 +27,12 @@ type Default interface {
 
 // Client exposes client APIs for the driver adapters
 type Client interface {
-	GetExample(orgID string, appID string, id string) (*model.Example, error)
-
 	// OccupationData APIs
 	GetOccupationData(code string) (*model.OccupationData, error)
 	GetAllOccupationDatas() ([]model.OccupationData, error)
 
 	// UserMatchingResult APIs
 	GetUserMatchingResult(id string) (*model.UserMatchingResult, error)
-	CreateUserMatchingResult(userMatchingResult model.UserMatchingResult) (*model.UserMatchingResult, error)
-	UpdateUserMatchingResult(userMatchingResult model.UserMatchingResult) error
 	DeleteUserMatchingResult(id string) error
 
 	// Survey Data APIs
@@ -44,34 +40,16 @@ type Client interface {
 	CreateSurveyData(surveyData model.SurveyData) (*model.SurveyData, error)
 	UpdateSurveyData(surveyData model.SurveyData) error
 	DeleteSurveyData(id string) error
+
+	// Occupation Matching
+	MatchOccupations(surveyData model.SurveyData, userID string)
 }
 
 // Admin exposes administrative APIs for the driver adapters
 type Admin interface {
-	GetExample(orgID string, appID string, id string) (*model.Example, error)
-	CreateExample(example model.Example) (*model.Example, error)
-	UpdateExample(example model.Example) error
-	AppendExample(example model.Example) (*model.Example, error)
-	DeleteExample(orgID string, appID string, id string) error
-
 	GetConfig(id string, claims *tokenauth.Claims) (*model.Config, error)
 	GetConfigs(configType *string, claims *tokenauth.Claims) ([]model.Config, error)
 	CreateConfig(config model.Config, claims *tokenauth.Claims) (*model.Config, error)
 	UpdateConfig(config model.Config, claims *tokenauth.Claims) error
 	DeleteConfig(id string, claims *tokenauth.Claims) error
-}
-
-// BBs exposes Building Block APIs for the driver adapters
-type BBs interface {
-	GetExample(orgID string, appID string, id string) (*model.Example, error)
-}
-
-// TPS exposes third-party service APIs for the driver adapters
-type TPS interface {
-	GetExample(orgID string, appID string, id string) (*model.Example, error)
-}
-
-// System exposes system administrative APIs for the driver adapters
-type System interface {
-	GetExample(orgID string, appID string, id string) (*model.Example, error)
 }