+++ core/interfaces/driven.go
@@ -30,17 +30,11 @@ type Storage interface {
 	UpdateConfig(config model.Config) error
 	DeleteConfig(id string) error
 
-	FindExample(orgID string, appID string, id string) (*model.Example, error)
-	InsertExample(example model.Example) error
-	UpdateExample(example model.Example) error
-	DeleteExample(orgID string, appID string, id string) error
-
 	GetOccupationData(id string) (*model.OccupationData, error)
 	GetAllOccupationDatas() ([]model.OccupationData, error)
 
 	GetUserMatchingResult(id string) (*model.UserMatchingResult, error)
-	CreateUserMatchingResult(bessiData model.UserMatchingResult) error
-	UpdateUserMatchingResult(bessiData model.UserMatchingResult) error
+	SaveUserMatchingResult(bessiData model.UserMatchingResult) error
 	DeleteUserMatchingResult(id string) error
 
 	GetSurveyData(id string) (*model.SurveyData, error)
@@ -52,5 +46,4 @@ type Storage interface {
 // StorageListener represents storage listener
 type StorageListener interface {
 	OnConfigsUpdated()
-	OnExamplesUpdated()
 }