+++ driver/web/apis_client.go
@@ -31,25 +31,6 @@ type ClientAPIsHandler struct {
 	app *core.Application
 }
 
-func (h ClientAPIsHandler) getExample(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
-	params := mux.Vars(r)
-	id := params["id"]
-	if len(id) <= 0 {
-		return l.HTTPResponseErrorData(logutils.StatusMissing, logutils.TypePathParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
-	}
-
-	surveyData, err := h.app.Client.GetExample(claims.OrgID, claims.AppID, id)
-	if err != nil {
-		return l.HTTPResponseErrorAction(logutils.ActionGet, model.TypeExample, nil, err, http.StatusInternalServerError, true)
-	}
-
-	response, err := json.Marshal(surveyData)
-	if err != nil {
-		return l.HTTPResponseErrorAction(logutils.ActionMarshal, logutils.TypeResponseBody, nil, err, http.StatusInternalServerError, false)
-	}
-	return l.HTTPResponseSuccessJSON(response)
-}
-
 func (h ClientAPIsHandler) getOccupationData(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
 	params := mux.Vars(r)
 	code := params["code"]
@@ -96,42 +77,6 @@ func (h ClientAPIsHandler) getUserMatchingResult(l *logs.Log, r *http.Request, c
 	return l.HTTPResponseSuccessJSON(response)
 }
 
-func (h ClientAPIsHandler) createUserMatchingResult(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
-	var requestData model.UserMatchingResult
-	err := json.NewDecoder(r.Body).Decode(&requestData)
-	if err != nil {
-		return l.HTTPResponseErrorAction(logutils.ActionUnmarshal, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
-	}
-	requestData.ID = claims.Subject
-	userMatchingResult, err := h.app.Client.CreateUserMatchingResult(requestData)
-	if err != nil || userMatchingResult == nil {
-		return l.HTTPResponseErrorAction(logutils.ActionCreate, model.TypeUserMatchingResult, nil, err, http.StatusInternalServerError, true)
-	}
-
-	response, err := json.Marshal(userMatchingResult)
-	if err != nil {
-		return l.HTTPResponseErrorAction(logutils.ActionMarshal, logutils.TypeResponseBody, nil, err, http.StatusInternalServerError, false)
-	}
-	return l.HTTPResponseSuccessJSON(response)
-}
-
-func (h ClientAPIsHandler) updateUserMatchingResult(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
-	var requestData model.UserMatchingResult
-	err := json.NewDecoder(r.Body).Decode(&requestData)
-	if err != nil {
-		return l.HTTPResponseErrorAction(logutils.ActionUnmarshal, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
-	}
-
-	id := claims.Subject
-	requestData.ID = id
-	err = h.app.Client.UpdateUserMatchingResult(requestData)
-	if err != nil {
-		return l.HTTPResponseErrorAction(logutils.ActionUpdate, model.TypeUserMatchingResult, nil, err, http.StatusInternalServerError, true)
-	}
-
-	return l.HTTPResponseSuccess()
-}
-
 func (h ClientAPIsHandler) deleteUserMatchingResult(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
 	id := claims.Subject
 	err := h.app.Client.DeleteUserMatchingResult(id)
@@ -169,10 +114,10 @@ func (h ClientAPIsHandler) createSurveyData(l *logs.Log, r *http.Request, claims
 	}
 
 	surveyData, err := h.app.Client.CreateSurveyData(requestData)
-	surveyData.ID = claims.Subject
 	if err != nil || surveyData == nil {
 		return l.HTTPResponseErrorAction(logutils.ActionCreate, model.TypeSurveyData, nil, err, http.StatusInternalServerError, true)
 	}
+	go h.app.Client.MatchOccupations(*surveyData, claims.Subject)
 
 	response, err := json.Marshal(surveyData)
 	if err != nil {