+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added endpoints for storing BESSI survey data [#1](https://github.com/ApoorvaAditya/skills-to-jobs-building-block/issues/1)
 - Added endpoints for storing user's matching results [#1](https://github.com/ApoorvaAditya/skills-to-jobs-building-block/issues/3)
 - Added endpoints for storing ONET occupation data [#4](https://github.com/ApoorvaAditya/skills-to-jobs-building-block/issues/4)
+- Added a matching algorithm for matching users to relevant occupations [#7](https://github.com/ApoorvaAditya/skills-to-jobs-building-block/issues/7)
 
 ### Fixed